requirements.txt
streamlit>=1.25.0
openai>=0.27.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
tavily-python>=0.3.0
pytest>=7.0.0